requirements.txt
langchain
python-dotenv
openai
google-api-core
uvicorn
fastapi
dnspython
alembic
pymysql
cryptography
pyjwt
